+++ apps/member-profile/app/routes/_profile.companies.tsx
@@ -19,7 +19,7 @@ import {
   ListCompaniesWhere,
 } from '@oyster/core/employment';
 import { listCompanies } from '@oyster/core/employment.server';
-import { track } from '@oyster/infrastructure/mixpanel';
+import { track } from '@oyster/core/mixpanel';
 import {
   cx,
   Dashboard,
