+++ apps/member-profile/app/routes/api.resources.$id.upvote.ts
@@ -1,7 +1,7 @@
 import { type ActionFunctionArgs, json } from '@remix-run/node';
 
 import { upvoteResource } from '@oyster/core/resources.server';
-import { track } from '@oyster/infrastructure/mixpanel';
+import { track } from '@oyster/core/mixpanel';
 
 import { ensureUserAuthenticated, user } from '@/shared/session.server';
