+++ apps/member-profile/app/routes/api.tags.add.ts
@@ -2,7 +2,7 @@ import { type ActionFunctionArgs, json } from '@remix-run/node';
 
 import { CreateTagInput } from '@oyster/core/resources';
 import { createTag } from '@oyster/core/resources.server';
-import { track } from '@oyster/infrastructure/mixpanel';
+import { track } from '@oyster/core/mixpanel';
 import { validateForm } from '@oyster/ui';
 
 import { ensureUserAuthenticated, user } from '@/shared/session.server';
