+++ apps/member-profile/app/shared/hooks/use-mixpanel-tracker.tsx
@@ -1,9 +1,6 @@
 import { useFetcher } from '@remix-run/react';
 
-import {
-  type MixpanelEvent,
-  type TrackInput,
-} from '@oyster/infrastructure/mixpanel';
+import { type MixpanelEvent, type TrackInput } from '@oyster/core/mixpanel';
 
 export function useMixpanelTracker() {
   const fetcher = useFetcher();
