+++ packages/core/src/modules/mixpanel/mixpanel.ts
