--- packages/infrastructure/.eslintrc.cjs
@@ -1,7 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  extends: ['@oyster/eslint-config/base'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-  },
-};
