--- packages/infrastructure/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "@oyster/tsconfig/base.json",
-  "include": ["src"],
-  "compilerOptions": {
-    "module": "ESNext",
-    "moduleResolution": "Bundler"
-  }
-}
